[C] Fix `snprintf` arguments.

+++ aeron-driver/src/main/c/aeron_driver_context.c
@@ -1407,13 +1407,16 @@ int aeron_driver_context_validate_mtu_length(size_t mtu_length)
         AERON_SET_ERR(
             EINVAL,
             "mtuLength must be a > HEADER_LENGTH and <= MAX_UDP_PAYLOAD_LENGTH: mtuLength=%" PRIu64,
-            mtu_length);
+            (uint64_t)mtu_length);
         return -1;
     }
 
     if ((mtu_length & (AERON_LOGBUFFER_FRAME_ALIGNMENT - 1)) != 0)
     {
-        AERON_SET_ERR(EINVAL, "mtuLength must be a multiple of FRAME_ALIGNMENT: mtuLength=%" PRIu64, mtu_length);
+        AERON_SET_ERR(
+            EINVAL,
+            "mtuLength must be a multiple of FRAME_ALIGNMENT: mtuLength=%" PRIu64,
+            (uint64_t)mtu_length);
         return -1;
     }
 
@@ -1428,9 +1431,9 @@ int aeron_driver_context_validate_receiver_window_length(const char* param, size
             EINVAL,
             "%s=%" PRIu64 " must be at least two times larger than the %s=%" PRIu64,
             param,
-            window_length,
+            (uint64_t)window_length,
             AERON_URI_MTU_LENGTH_KEY,
-            mtu_length);
+            (uint64_t)mtu_length);
         return -1;
     }
     return 0;
@@ -1451,9 +1454,9 @@ int aeron_driver_context_validate_publisher_window_length(
             EINVAL,
             "%s=%" PRIu64 " must not exceed half the %s=%" PRIu64,
             param,
-            window_length,
+            (uint64_t)window_length,
             AERON_URI_TERM_LENGTH_KEY,
-            term_length);
+            (uint64_t)term_length);
         return -1;
     }
     return 0;
